backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
dnspython
python-dotenv
pydantic[email]
python-multipart
dotenv
pymongo
